Split rules up into ORM and non-ORM flavours.
Having a RulesChecker independent of the ORM will be very helpful when
implementing rules on the elastic search ODM.

This should be fully backwards compatible as the old classes and methods
are still around.
